LayoutTests/fast/events/touch/ios/touch-event-regions/overflow-expected.txt

touch start on container
touchstart inside overflow
touchstart
58, 50 - 260, 252
69, 453 - 191, 555
touchstart

touchend

touchforcechange

passive
